changed first screen background color to blue. Also in purchase changed the id to long int

+++ purchase.cpp
@@ -1,7 +1,7 @@
 #include "purchase.h"
 #include <QtWidgets/QCalendarWidget>
 
-Purchase::Purchase(int id,QString categ,double amount,QString desc,QDate date,QString people,QString paym,QString place,QString event)
+Purchase::Purchase(long int id,QString categ,double amount,QString desc,QDate date,QString people,QString paym,QString place,QString event)
 {
     this->category=categ;
     this->amount=amount;
