+++ purchase.h
@@ -19,7 +19,7 @@ private:
     QString event;
     int id;
 public:
-        Purchase(int id,QString categ,double amount,QString desc,QDate date,QString people,QString paym,QString place,QString event);
+        Purchase(long int id,QString categ,double amount,QString desc,QDate date,QString people,QString paym,QString place,QString event);
         Purchase();
 
         QString getCategory() { return category;}
